// frontend/src/types/gorm-models.ts
/**
 * This file is automatically generated by frontend/scripts/generate-types.js at build time.
 * DO NOT EDIT MANUALLY! Any changes will be overwritten during building or running.
 */

export interface GormAccount {
  id: number
  employee_id: number | null
  username: string | null
  password: string | null
}

export interface GormEmployee {
  id: number
  employee_number: string | null
  name: string | null
  department: string | null
}

export interface GormQuoteEmployee {
  id: number
  employee_id: number | null
  role: string | null
}

export const DEFAULT_GORM_ACCOUNT: Omit<GormAccount, 'password'> = {
  id: 0,
  employee_id: null,
  username: null,
}

export const DEFAULT_GORM_QUOTE_EMPLOYEE: GormQuoteEmployee = {
  id: 0,
  employee_id: null,
  role: null,
}

export interface User extends GormEmployee {
  account: Omit<GormAccount, 'password'> | null
  quote_employee: GormQuoteEmployee | null
}

export const DEFAULT_USER_NULL: User = {
  id: 0,
  employee_number: null,
  name: null,
  department: null,
  account: null,
  quote_employee: null,
}